Change the type of SparseMatrix::num_nonzeroes and SparseMatrix::max_num_nonzeroes from int to int64_t.
This will allow for larger size Jacobians to be allocated, when memory
is available and matrix dimensions are still 32bit.

https://github.com/ceres-solver/ceres-solver/issues/412

+++ internal/ceres/block_random_access_diagonal_matrix.cc
@@ -55,7 +55,7 @@ BlockRandomAccessDiagonalMatrix::BlockRandomAccessDiagonalMatrix(
   // Build the row/column layout vector and count the number of scalar
   // rows/columns.
   int num_cols = 0;
-  int num_nonzeros = 0;
+  int64_t num_nonzeros = 0;
   vector<int> block_positions;
   for (int i = 0; i < blocks_.size(); ++i) {
     block_positions.push_back(num_cols);
